Fixing issues detected by shellcheck

+++ systembackup.sh
@@ -35,10 +35,7 @@ source /etc/backup/systembackup.conf
 CURRENT_TIME=$(date -Ihours)
 
 # Last backup date
-LOCAL_LAST_BACKUP=$(ls $LOCAL_BACKUP_DIRECTORY | tail -n 1)
-
-# Link directory
-LOCAL_BACKUP_LINK="--link-dest=$LOCAL_BACKUP_DIRECTORY/$LOCAL_LAST_BACKUP/"
+LOCAL_LAST_BACKUP=$(ls "$LOCAL_BACKUP_DIRECTORY" | tail -n 1)
 
 # Backup location
 LOCAL_BACKUP_TARGET="$LOCAL_BACKUP_DIRECTORY/$CURRENT_TIME"
@@ -67,7 +64,7 @@ do
   IFS=$'\n'
   for j in $(find "$i" -type f)
   do
-    #echo "working on file $j"
+    echo "working on file $j"
 
     LAST_HMAC="$LOCAL_BACKUP_DIRECTORY/$LOCAL_LAST_BACKUP$j.hmac$HMAC_ALGO"
     LAST_ENC="$LOCAL_BACKUP_DIRECTORY/$LOCAL_LAST_BACKUP$j.enc$CIPHER"
@@ -82,17 +79,16 @@ do
       # Should also be nice on memory usage too!
 
       # Check that the hmac is valid, THEN check if the crypto needs to be done.
-      #echo "cmp 1"
-      cmp -s "$LAST_HMAC" <(cat "$LAST_ENC" | openssl dgst "$HMAC_ALGO" -hmac "$HMAC_KEY" -r | cut -f 1 -d " ") 
-      if [ $? -eq 0 ]
+      echo "cmp 1"
+      if cmp -s "$LAST_HMAC" <(openssl dgst "$HMAC_ALGO" -hmac "$HMAC_KEY" -r < $"LAST_ENC" | cut -f 1 -d " ")
       then
         # This does some decryption, because that is what it needs to get the IV.
-        SALT_IV="$(openssl enc -d $CIPHER -pass $PASSWORD -P -in "$LAST_ENC")"
-        SALT="$(echo $SALT_IV | cut -d " " -f 1 | cut -d '=' -f 2)"
-        IV="$(echo $SALT_IV | cut -d " " -f 4 | cut -d '=' -f 2)"
-        #echo "salt = $SALT"
-        #echo "iv = $IV"
-        #echo "cmp 2"
+        SALT_IV="$(openssl enc -d "$CIPHER" -pass "$PASSWORD" -P -in "$LAST_ENC")"
+        SALT="$(echo "$SALT_IV" | cut -d " " -f 1 | cut -d '=' -f 2)"
+        IV="$(echo "$SALT_IV" | cut -d " " -f 4 | cut -d '=' -f 2)"
+        echo "salt = $SALT"
+        echo "iv = $IV"
+        echo "cmp 2"
         cmp -s "$LAST_ENC" <(openssl enc -e "$CIPHER" -pass "$PASSWORD" -S "$SALT" -iv "$IV" -in "$j")
         status=$?
       fi
@@ -100,19 +96,19 @@ do
       then
         # Files are the same
         #echo "status = $status"
-        #echo "Files are the same"
+        echo "Files are the same"
         ln "$LAST_ENC" "$TARGET_ENC"
         ln "$LAST_HMAC" "$TARGET_HMAC"
       else
         # Files are differnet
         #echo "status = $status"
-        #echo "Files are different"
+        echo "Files are different"
 
         # Write encrypted file to disk and generate hmac
         openssl enc -e "$CIPHER" -pass "$PASSWORD" -in "$j" | tee "$TARGET_ENC" | openssl dgst "$HMAC_ALGO" -hmac "$HMAC_KEY" -r | cut -f 1 -d " " > "$TARGET_HMAC" 
       fi
     else
-      #echo "new file, encrypt it."
+      echo "new file, encrypt it."
       openssl enc -e "$CIPHER" -pass "$PASSWORD" -in "$j" | tee "$TARGET_ENC" | openssl dgst "$HMAC_ALGO" -hmac "$HMAC_KEY" -r | cut -f 1 -d " " > "$TARGET_HMAC"
     fi
   done
@@ -145,17 +141,17 @@ fi
 pacman -Qqem | openssl enc -e "$CIPHER" -pass "$PASSWORD" | tee "$LOCAL_BACKUP_TARGET/pkglist_aur.txt.enc$CIPHER" | openssl dgst "$HMAC_ALGO" -hmac "$HMAC_KEY" -r | cut -f 1 -d " " > "$LOCAL_BACKUP_TARGET/pkglist_aur.txt.hmac$HMAC_ALGO"
 
 # Only keep a limited number of backups
-while [ $(ls $LOCAL_BACKUP_DIRECTORY | wc -l) -gt $LOCAL_BACKUPS_TO_KEEP ]
+while [ "$(ls "$LOCAL_BACKUP_DIRECTORY" | wc -l)" -gt "$LOCAL_BACKUPS_TO_KEEP" ]
 do
-  echo "deleting $LOCAL_BACKUP_DIRECTORY/$(ls $LOCAL_BACKUP_DIRECTORY | head -1)"
-  rm -rf "$LOCAL_BACKUP_DIRECTORY/$(ls $LOCAL_BACKUP_DIRECTORY | head  -1)"
+  echo "deleting $LOCAL_BACKUP_DIRECTORY/$(ls "$LOCAL_BACKUP_DIRECTORY" | head -1)"
+  rm -rf "${LOCAL_BACKUP_DIRECTORY:?}/$(ls "$LOCAL_BACKUP_DIRECTORY" | head  -1)"
 done
 
 # Run remote backups
 for REMOTE_CONFIG_LINE in "${REMOTE_CONFIG[@]}"
 do
   # Split the config line
-  IFS=' ' read -a SPLIT_REMOTE_CONFIG <<< $REMOTE_CONFIG_LINE
+  IFS=' ' read -ar SPLIT_REMOTE_CONFIG <<< "$REMOTE_CONFIG_LINE"
   REMOTE_SERVER=${SPLIT_REMOTE_CONFIG[0]}
   REMOTE_USER=${SPLIT_REMOTE_CONFIG[1]}
   REMOTE_SSH_ID=${SPLIT_REMOTE_CONFIG[2]}
@@ -165,27 +161,27 @@ do
   # Copy the backup to the remote server.
   # This does not create a new backup, 
   # but used the backup that was just made.
-  REMOTE_LAST_BACKUP=$(ssh -xo "BatchMode yes" -i $REMOTE_SSH_ID $REMOTE_USER@$REMOTE_SERVER ls $REMOTE_BACKUP_DIRECTORY | tail -n 1)
+  REMOTE_LAST_BACKUP=$(ssh -xo "BatchMode yes" -i "$REMOTE_SSH_ID" "$REMOTE_USER@$REMOTE_SERVER" ls "$REMOTE_BACKUP_DIRECTORY" | tail -n 1)
   REMOTE_BACKUP_TARGET="$REMOTE_BACKUP_DIRECTORY/$CURRENT_TIME"
   # Check if we need to do a full backup, or an update
-  if [ $REMOTE_LAST_BACKUP != $CURRENT_TIME  ]
+  if [ "$REMOTE_LAST_BACKUP" != "$CURRENT_TIME"  ]
   then
     echo "remote directory cp --reflink=auto -rp $REMOTE_BACKUP_DIRECTORY/$REMOTE_LAST_BACKUP $REMOTE_BACKUP_TARGET"
-    ssh -xo "BatchMode yes" -i $REMOTE_SSH_ID $REMOTE_USER@$REMOTE_SERVER sudo cp --reflink=auto -rp $REMOTE_BACKUP_DIRECTORY/$REMOTE_LAST_BACKUP $REMOTE_BACKUP_TARGET
-    echo "remote rsync -ze ssh -i $REMOTE_SSH_ID $RYSNC_OPTIONS --inplace $LOCAL_BACKUP_TARGET $REMOTE_USER@$REMOTE_SERVER:$REMOTE_BACKUP_DIRECTORY"
-    rsync -ze "ssh -xi $REMOTE_SSH_ID" $RSYNC_OPTIONS --inplace $LOCAL_BACKUP_TARGET $REMOTE_USER@$REMOTE_SERVER:$REMOTE_BACKUP_DIRECTORY 
+    ssh -xo "BatchMode yes" -i "$REMOTE_SSH_ID" "$REMOTE_USER@$REMOTE_SERVER" sudo cp --reflink=auto -rp "$REMOTE_BACKUP_DIRECTORY/$REMOTE_LAST_BACKUP $REMOTE_BACKUP_TARGET"
+    echo "remote rsync -ze ssh -i $REMOTE_SSH_ID $RSYNC_OPTIONS --inplace $LOCAL_BACKUP_TARGET $REMOTE_USER@$REMOTE_SERVER:$REMOTE_BACKUP_DIRECTORY"
+    rsync -ze "ssh -xi $REMOTE_SSH_ID" "$RSYNC_OPTIONS" --inplace "$LOCAL_BACKUP_TARGET" "$REMOTE_USER@$REMOTE_SERVER:$REMOTE_BACKUP_DIRECTORY"
   else
     echo "remote rsync -ze ssh -i $REMOTE_SSH_ID $RSYNC_OPTIONS $LOCAL_BACKUP_TARGET $REMOTE_USER@$REMOTE_SERVER:$REMOTE_BACKUP_DIRECTORY/$REMOTE_LAST_BACKUP"
-    rsync -ze "ssh -i $REMOTE_SSH_ID" $RSYNC_OPTIONS $LOCAL_BACKUP_TARGET $REMOTE_USER@$REMOTE_SERVER:$REMOTE_BACKUP_DIRECTORY/$REMOTE_LAST_BACKUP 
+    rsync -ze "ssh -i $REMOTE_SSH_ID" "$RSYNC_OPTIONS" "$LOCAL_BACKUP_TARGET" "$REMOTE_USER@$REMOTE_SERVER:$REMOTE_BACKUP_DIRECTORY/$REMOTE_LAST_BACKUP"
   fi
 
   # Only keep a limited number of backups 
   # on the remote server
-  while [ $(ssh -xo "BatchMode yes" -i $REMOTE_SSH_ID $REMOTE_USER@$REMOTE_SERVER ls $REMOTE_BACKUP_DIRECTORY | wc -l) -gt $REMOTE_BACKUPS_TO_KEEP ]
+  while [ "$(ssh -xo "BatchMode yes" -i "$REMOTE_SSH_ID" "$REMOTE_USER@$REMOTE_SERVER" ls "$REMOTE_BACKUP_DIRECTORY" | wc -l)" -gt "$REMOTE_BACKUPS_TO_KEEP" ]
   do
-    REMOTE_DIRECTORY_TO_DELETE=$REMOTE_BACKUP_DIRECTORY/$(ssh -xo "BatchMode yes" -i $REMOTE_SSH_ID $REMOTE_USER@$REMOTE_SERVER ls $REMOTE_BACKUP_DIRECTORY | head -1)
+    REMOTE_DIRECTORY_TO_DELETE="$REMOTE_BACKUP_DIRECTORY/$(ssh -xo "BatchMode yes" -i "$REMOTE_SSH_ID" "$REMOTE_USER@$REMOTE_SERVER" ls "$REMOTE_BACKUP_DIRECTORY" | head -1)"
     echo "deleting $REMOTE_USER@$REMOTE_SERVER:$REMOTE_DIRECTORY_TO_DELETE"
-    ssh -xo "BatchMode yes" -i $REMOTE_SSH_ID $REMOTE_USER@$REMOTE_SERVER sudo rm -rf $REMOTE_DIRECTORY_TO_DELETE
+    ssh -xo "BatchMode yes" -i "$REMOTE_SSH_ID" "$REMOTE_USER@$REMOTE_SERVER" sudo rm -rf "$REMOTE_DIRECTORY_TO_DELETE"
   done
 done
 echo "Backup completed"
